packages/backend/pyproject.toml
[tool.poetry]
name = "ai-flow-back"
version = "0.11.3"
description = ""
authors = ["DahnM20 <you@example.com>"]
readme = "README.md"
packages = [{ include = "app" }]


[tool.poetry.dependencies]
python = ">=3.11,<3.12"
python-dotenv = "^1.0.0"
openai = "1.76.0"
flask = "^2.2.3"
flask-socketio = "^5.3.3"
flask-cors = "^3.0.10"
unstructured = "^0.6.3"
langchain = ">=0.0.303"
python-magic = "^0.4.14"
pytesseract = "^0.3.10"
requests = "^2.31.0"
tabulate = "^0.9.0"
pdf2image = "^1.16.3"
colorlog = "^6.7.0"
eventlet = "^0.33.3"
playwright = "1.39"
youtube-transcript-api = "^0.6.1"
pytube = "^15.0.0"
boto3 = "^1.28.52"
pyjwt = "^2.8.0"
jwcrypto = "^1.5.0"
python-jose = "^3.3.0"
cachetools = "^5.3.1"
flask-injector = "^0.15.0"
setuptools = "^68.2.2"
pypdf = "4.2.0"
replicate = "0.22.0"
tiktoken = "0.7.0"
openapi-spec-validator = "^0.7.1"
anthropic = "0.49.0"
pymupdf = "^1.24.7"
pydub = "^0.25.1"
markdownify = "^1.1.0"
beautifulsoup4 = "^4.13.3"
ultralytics = "^8.0.0"
opencv-python = "^4.5.0"
Pillow = "^10.0.0"
pymodbus = "^3.11.1"
numpy = "^2.3.2"
langchain-community = "^0.3.29"

[tool.poetry.group.dev.dependencies]
datamodel-code-generator = "^0.25.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
